+++ R/defaults.R
@@ -155,6 +155,12 @@ set_flextable_defaults <- function(
   if( !is.null(extra_css) ){
     x$extra_css <- extra_css
   }
+  if( !is.null(fmt_date) ){
+    x$fmt_date <- fmt_date
+  }
+  if( !is.null(fmt_datetime) ){
+    x$fmt_datetime <- fmt_datetime
+  }
 
   if( !is.null(theme_fun) && is.character(theme_fun) && length(theme_fun) == 1 ){
     x$theme_fun <- theme_fun
@@ -197,6 +203,110 @@ set_flextable_defaults <- function(
 #' get_flextable_defaults()
 #' @family functions related to themes
 get_flextable_defaults <- function(){
-  flextable_global$defaults
+  x <- flextable_global$defaults
+  class(x) <- "flextable_defaults"
+  x
+}
+
+#' @export
+print.flextable_defaults <- function(x, ...){
+
+  message("## style properties\n")
+  styles <- c("font.family", "font.size", "font.color", "text.align", "padding.bottom",
+    "padding.top", "padding.left", "padding.right", "border.color",
+    "background.color")
+  df <- data.frame(property = styles, value = unlist(x[styles]), stringsAsFactors = FALSE)
+  row.names(df) <- NULL
+  print(df)
+  message("")
+
+  message("## cell content settings\n")
+  contents <- c("decimal.mark", "big.mark",
+              "digits", "na_str", "fmt_date", "fmt_datetime")
+  df <- data.frame(property = contents, value = unlist(x[contents]), stringsAsFactors = FALSE)
+  row.names(df) <- NULL
+  print(df)
+  message("")
+
+  message("## table.layout is:", x$table.layout, "\n")
+  message("## default theme is:", x$theme_fun, "\n")
+
+  message("## HTML specific:")
+  message("extra_css:", x$extra_css)
+  message("post_process_html:")
+  print(x$post_process_html)
+  message("")
+
+  message("## latex specific:")
+  message("post_process_pdf:")
+  print(x$post_process_pdf)
+  message("")
+
+  message("## Word specific:")
+  message("post_process_docx:")
+  print(x$post_process_docx)
+  message("")
+
+  message("## PowerPoint specific:")
+  message("post_process_pptx:")
+  print(x$post_process_pptx)
+  message("")
+
+  invisible(NULL)
+}
+
+
+
+
+#' @export
+#' @inheritParams officer::fp_text
+#' @title Text formatting properties
+#'
+#' @description Create a [fp_text()] object that uses
+#' defaut values defined in flextable defaults formatting properties, i.e.
+#' default font color, font size and font family (see [set_flextable_defaults()]).
+#' (see [set_flextable_defaults()]).
+#' @examples
+#' fp_text_default(bold = TRUE)
+#' @family functions for defining formatting properties
+#' @seealso [as_chunk()]
+fp_text_default <- function(color = flextable_global$defaults$font.color,
+                            font.size = flextable_global$defaults$font.size,
+                            bold = FALSE,
+                            italic = FALSE,
+                            underlined = FALSE,
+                            font.family = flextable_global$defaults$font.family,
+                            vertical.align = "baseline",
+                            shading.color = "transparent"){
+  fp_text(
+    color = color,
+    font.size = font.size,
+    bold = bold,
+    italic = italic,
+    underlined = underlined,
+    font.family = font.family,
+    vertical.align = vertical.align,
+    shading.color = shading.color
+  )
+
+}
+#' @export
+#' @inheritParams officer::fp_border
+#' @title Border formatting properties
+#'
+#' @description Create a [fp_border()] object that uses
+#' defaut values defined in flextable defaults formatting properties, i.e.
+#' default border color (see [set_flextable_defaults()]).
+#' @examples
+#' fp_border_default(width = 2)
+#' @family functions for defining formatting properties
+#' @seealso [hline()], [vline()]
+fp_border_default <- function(color = flextable_global$defaults$border.color,
+                              style = "solid", width = 1){
+  fp_border(
+    color = color,
+    style = style,
+    width = width)
+
 }